+++ distrib/deb/create_data_package.sh
@@ -1,8 +1,11 @@
 #!/bin/bash
 
+if [ $USER != root ] ; then echo This program should be run from root account - type \"sudo $0\" ; exit 1 ; fi
 echo Creating openlierox-data.deb
 if [ -e pkg ] ; then rm -rf pkg ; fi
 mkdir -p pkg/usr/share
+cp -r data/share pkg/usr
+gzip -c -9 ../../doc/ChangeLog > pkg/usr/share/doc/openlierox-data/changelog.gz
 if [ -e ../../share/gamedir/front.bmp ] ; then rm ../../share/gamedir/front.bmp ; fi
 if [ -e ../../share/gamedir/back.bmp ] ; then rm ../../share/gamedir/back.bmp ; fi
 
@@ -15,8 +18,9 @@ find pkg/ -type "f" -exec md5sum {} \; | sed 's@ pkg/@ @' > md5sums
 mkdir pkg/DEBIAN
 mv md5sums pkg/DEBIAN
 find data -maxdepth 1 -type "f" -exec cp {} pkg/DEBIAN \;
-find pkg/ -name ".svn" -exec rm -rf {} \; >/dev/null 2>&1
-find pkg/ -name "*~" -exec rm -rf {} \; >/dev/null 2>&1
+
+chgrp -R root pkg/
+chown -R root pkg/
 
 Version=`cat ../../VERSION | sed 's/_/./g'`
 Size=`du -k -s pkg | grep -o '[0-9]*'`
@@ -27,4 +31,4 @@ sed "s/^Installed-Size:.*/Installed-Size: $Size/" \
 
 dpkg -b pkg openlierox-data_$Version.deb
 
-# rm -rf pkg
+rm -rf pkg
